Update Note class to Notes class to support multiple notes
